Change from TypeScript@4.1.0 to TypeScript@3.3.4000
Due to issues related to styled-components being slow in the latest
stable version of TypeScript(4.1.0), the version of this package had to
be changed.

Also, a change was made in the src/styles/breakpoints.ts file regarding
the fact that, in v3.3.4000, TypeScript didn't consider ReadOnlyArray,
such as TemplateStringsArray, as a valid type for rest parameters.