Change posts structure and images things
In order to have a well defined structure for the posts, each post now
has its own folder. All the images related to a post will be under this
post's folder. Besides, both the post content in markdown and the cover
image now have a specific name. With that, it will no longer be
necessary to specify the cover image in the post's markdown frontmatter.

All the images in a post, including the cover, now have the blur up
effect. Alongside that, each image now uses the srcset and sizes
attributes, allowing the users to have a better mobile experience.
Because of that, some images had to be resized.

It was also added a new field to the MarkdownRemark nodes. That field
will be used when the pages start to be written in markdown.