Fix responsiveImgMediaQueries in egen.yaml
The previous value representing the margin of the image used in this
field (2rem) was wrong for two reasons:

- The correct margin value is 40px (4rem) because it's 2rem on either
side.
- According to my tests, using rem units in the img tag's sizes
attribute is not reliable. This happens because sometimes the root's new
font-size value (62.5%), which makes 1rem=10px, hasn't already been
applied by the browser.

Because of these reasons, the img's margin value present in the
responsiveImgMediaQueries field was updated to 2.5em. This way, the value
continues to be relative to the user's font-size value, but in a more
reliable way.